vue-dialog(refactor): remove useDialog
