+++ packages/dialog/vue/lib/main.ts
@@ -19,8 +19,3 @@ export {default as DialogDescription} from './DialogDescription.vue';
 export {default as DialogRoot} from './DialogRoot.vue';
 export {default as DialogTitle} from './DialogTitle.vue';
 export {default as DialogTrigger} from './DialogTrigger.vue';
-export {
-	useDialog,
-	type UseDialogOptions,
-	type UseDialogValue,
-} from './useDialog';
